refact: moving setting components files from component/ to lib/
