+++ apps/web/lib/settings/index.ts
@@ -1,0 +1,13 @@
+export * from './left-side-setting-menu';
+export * from './danger-zone-personal';
+export * from './danger-zone-team';
+export * from './language-dropdown';
+export * from './list-card';
+export * from './personal-setting-form';
+export * from './profile-avatar';
+export * from './task-labels-form';
+export * from './task-priorities-form';
+export * from './task-sizes-form';
+export * from './task-statuses-form';
+export * from './team-setting-form';
+export * from './timezone-dropdown';
